+++ worker/workerpool_test.go
@@ -2,66 +2,14 @@ package worker
 
 import (
 	"context"
-	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
-
-	"golang.org/x/time/rate"
 )
 
 // getActiveWorkers is a test helper to safely read activeWorkers.
 func (wp *WorkerPool) getActiveWorkers() int64 {
 	return wp.activeWorkers.Load()
-}
-
-// getTargetConcurrency is a test helper to safely read targetConcurrency.
-func (wp *WorkerPool) getTargetConcurrency() int64 {
-	return wp.targetConcurrency.Load()
-}
-
-func TestSetConcurrency(t *testing.T) {
-	// Test the low-level SetConcurrency API directly.
-	wp := NewMultiStepWorkerPool(
-		func(ctx context.Context, wp *WorkerPool) error {
-			time.Sleep(10 * time.Millisecond)
-			return nil
-		},
-		[]*Step{
-			NewStep(
-				WithDuration(0), // Infinite duration, so runProfileSteps won't overwrite concurrency.
-				WithConcurrency(1),
-			),
-		},
-	)
-
-	wp.Launch()
-	defer wp.Stop()
-
-	// Wait for step to start workers.
-	time.Sleep(150 * time.Millisecond)
-
-	// Check initial state.
-	if got := wp.getActiveWorkers(); got != 1 {
-		t.Errorf("expected 1 worker, got %d", got)
-	}
-	if got := wp.getTargetConcurrency(); got != 1 {
-		t.Errorf("expected concurrency 1, got %d", got)
-	}
-
-	// Increase to 5 workers.
-	wp.SetConcurrency(5)
-	time.Sleep(50 * time.Millisecond)
-	if got := wp.getActiveWorkers(); got != 5 {
-		t.Errorf("expected 5 workers, got %d", got)
-	}
-
-	// Decrease to 2 workers.
-	wp.SetConcurrency(2)
-	time.Sleep(50 * time.Millisecond)
-	if got := wp.getActiveWorkers(); got != 2 {
-		t.Errorf("expected 2 workers, got %d", got)
-	}
 }
 
 func TestSetConcurrencyAfterWorkersTerminate(t *testing.T) {
@@ -94,28 +42,6 @@ func TestSetConcurrencyAfterWorkersTerminate(t *testing.T) {
 	}
 }
 
-func TestRunProfile(t *testing.T) {
-	wp := NewWorkerPool(
-		func(ctx context.Context, wp *WorkerPool) error {
-			return nil
-		},
-		WithDuration(100*time.Millisecond),
-		WithRateLimit(20, 20),
-		WithConcurrency(2),
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	// Check final state
-	if got := wp.getTargetConcurrency(); got != 2 {
-		t.Errorf("expected final concurrency 2, got %d", got)
-	}
-	if wp.limiter.Limit() != rate.Limit(20) {
-		t.Errorf("expected final RPS 20, got %v", wp.limiter.Limit())
-	}
-}
-
 func TestRateLimitWithEasing(t *testing.T) {
 	st := NewStep(
 		WithDuration(time.Second),
@@ -128,267 +54,6 @@ func TestRateLimitWithEasing(t *testing.T) {
 	}
 	if st.concurrencyEasing(0.5) != 0.75 { // EaseOut(0.5) = 0.5*(2-0.5) = 0.75.
 		t.Errorf("expected easingConcurrency(0.5) to be 0.75, got %v", st.concurrencyEasing(0.5))
-	}
-}
-
-func TestMultiStepWorkerPool(t *testing.T) {
-	callCount := int32(0)
-
-	worker := func(ctx context.Context, wp *WorkerPool) error {
-		atomic.AddInt32(&callCount, 1)
-		return nil
-	}
-
-	steps := []*Step{
-		NewStep(
-			WithDuration(50*time.Millisecond),
-			WithConcurrency(2),
-			WithRateLimit(100, 100),
-		),
-		NewStep(
-			WithDuration(50*time.Millisecond),
-			WithConcurrency(4),
-			WithRateLimit(200, 200),
-		),
-	}
-
-	wp := NewMultiStepWorkerPool(worker, steps)
-	wp.Launch()
-	wp.Wait()
-
-	// Check final state.
-	if got := wp.getTargetConcurrency(); got != 4 {
-		t.Errorf("expected final concurrency 4, got %d", got)
-	}
-	if wp.limiter.Limit() != rate.Limit(200) {
-		t.Errorf("expected final RPS 200, got %v", wp.limiter.Limit())
-	}
-	if atomic.LoadInt32(&callCount) == 0 {
-		t.Error("expected worker to be called at least once")
-	}
-}
-
-func TestGlobalRepetitions(t *testing.T) {
-	var callCount int64
-
-	worker := func(ctx context.Context, wp *WorkerPool) error {
-		atomic.AddInt64(&callCount, 1)
-		return nil
-	}
-
-	wp := NewWorkerPool(
-		worker,
-		WithRepetitions(50),
-		WithConcurrency(10),
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	got := atomic.LoadInt64(&callCount)
-	if got != 50 {
-		t.Errorf("expected exactly 50 calls with global repetitions, got %d", got)
-	}
-}
-
-func TestRepetitionsAndDuration(t *testing.T) {
-	// Case 1: Repetitions run out before Duration.
-	t.Run("RepetitionsPrecedence", func(t *testing.T) {
-		var callCount int64
-		wp := NewWorkerPool(
-			func(ctx context.Context, wp *WorkerPool) error {
-				atomic.AddInt64(&callCount, 1)
-				return nil
-			},
-			WithDuration(5*time.Second), // Long duration.
-			WithRepetitions(10),         // Only 10 reps.
-			WithConcurrency(2),
-		)
-
-		start := time.Now()
-		wp.Launch()
-		wp.Wait()
-		elapsed := time.Since(start)
-
-		if elapsed > 2*time.Second {
-			t.Errorf("expected to finish quickly due to repetition limit, took %v", elapsed)
-		}
-
-		got := atomic.LoadInt64(&callCount)
-		if got != 10 {
-			t.Errorf("expected 10 calls, got %d", got)
-		}
-	})
-
-	// Case 2: Duration runs out before Repetitions.
-	t.Run("DurationPrecedence", func(t *testing.T) {
-		var callCount int64
-		wp := NewWorkerPool(
-			func(ctx context.Context, wp *WorkerPool) error {
-				atomic.AddInt64(&callCount, 1)
-				// Small sleep to ensure we don't accidentally burn through reps very fast (though 10M is safe).
-				time.Sleep(1 * time.Microsecond)
-				return nil
-			},
-			WithDuration(200*time.Millisecond),
-			WithRepetitions(1000000), // Huge number.
-			WithConcurrency(10),
-		)
-
-		start := time.Now()
-		wp.Launch()
-		wp.Wait()
-		elapsed := time.Since(start)
-
-		// Allow some slack in timing.
-		if elapsed < 200*time.Millisecond {
-			t.Errorf("finished too early: %v", elapsed)
-		}
-
-		got := atomic.LoadInt64(&callCount)
-		if got < 100 {
-			t.Errorf("expected some work done, got %d", got)
-		}
-		if got >= 1000000 {
-			t.Errorf("should not have finished all repetitions")
-		}
-	})
-}
-
-func TestWithInfiniteRepetitions(t *testing.T) {
-	var callCount int64
-
-	wp := NewWorkerPool(
-		func(ctx context.Context, wp *WorkerPool) error {
-			atomic.AddInt64(&callCount, 1)
-			return nil
-		},
-		WithInfiniteRepetitions(),
-		WithDuration(100*time.Millisecond),
-		WithConcurrency(2),
-		WithRateLimit(100, 100),
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	got := atomic.LoadInt64(&callCount)
-	// With infinite repetitions and 100ms duration at ~100 RPS, should have ~10+ calls
-	if got < 5 {
-		t.Errorf("expected at least 5 calls with infinite repetitions, got %d", got)
-	}
-}
-
-func TestWithWorkerFunc(t *testing.T) {
-	var defaultCalls, overrideCalls int64
-
-	defaultWorker := func(ctx context.Context, wp *WorkerPool) error {
-		atomic.AddInt64(&defaultCalls, 1)
-		return nil
-	}
-
-	overrideWorker := func(ctx context.Context, wp *WorkerPool) error {
-		atomic.AddInt64(&overrideCalls, 1)
-		return nil
-	}
-
-	steps := []*Step{
-		NewStep(
-			WithDuration(50*time.Millisecond),
-			WithConcurrency(1),
-			WithRateLimit(50, 50),
-		),
-		NewStep(
-			WithDuration(50*time.Millisecond),
-			WithConcurrency(1),
-			WithRateLimit(50, 50),
-			WithWorkerFunc(overrideWorker),
-		),
-	}
-
-	wp := NewMultiStepWorkerPool(defaultWorker, steps)
-	wp.Launch()
-	wp.Wait()
-
-	if atomic.LoadInt64(&defaultCalls) == 0 {
-		t.Error("expected default worker to be called in first step")
-	}
-	if atomic.LoadInt64(&overrideCalls) == 0 {
-		t.Error("expected override worker to be called in second step")
-	}
-}
-
-func TestWithHooks(t *testing.T) {
-	var onStartCalled, onEndCalled bool
-
-	onStart := func(ctx context.Context, wp *WorkerPool) {
-		onStartCalled = true
-	}
-
-	onEnd := func(ctx context.Context, wp *WorkerPool) {
-		onEndCalled = true
-	}
-
-	wp := NewWorkerPool(
-		func(ctx context.Context, wp *WorkerPool) error {
-			return nil
-		},
-		WithDuration(50*time.Millisecond),
-		WithConcurrency(1),
-		WithHooks(onStart, onEnd),
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	if !onStartCalled {
-		t.Error("expected onStart hook to be called")
-	}
-	if !onEndCalled {
-		t.Error("expected onEnd hook to be called")
-	}
-}
-
-func TestSetWorker(t *testing.T) {
-	var worker1Calls, worker2Calls int64
-
-	worker1 := func(ctx context.Context, wp *WorkerPool) error {
-		atomic.AddInt64(&worker1Calls, 1)
-		time.Sleep(10 * time.Millisecond)
-		return nil
-	}
-
-	worker2 := func(ctx context.Context, wp *WorkerPool) error {
-		atomic.AddInt64(&worker2Calls, 1)
-		time.Sleep(10 * time.Millisecond)
-		return nil
-	}
-
-	wp := NewMultiStepWorkerPool(
-		worker1,
-		[]*Step{
-			NewStep(
-				WithDuration(0),
-				WithConcurrency(1),
-			),
-		},
-	)
-
-	wp.Launch()
-	time.Sleep(50 * time.Millisecond)
-
-	// Change the worker function
-	wp.SetWorker(worker2)
-	time.Sleep(50 * time.Millisecond)
-
-	wp.Stop()
-	wp.Wait()
-
-	if atomic.LoadInt64(&worker1Calls) == 0 {
-		t.Error("expected worker1 to be called")
-	}
-	if atomic.LoadInt64(&worker2Calls) == 0 {
-		t.Error("expected worker2 to be called after SetWorker")
 	}
 }
 
@@ -492,46 +157,6 @@ func TestSetConcurrencyNegative(t *testing.T) {
 	wp.Wait()
 }
 
-func TestLaunchWithContextCancellation(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-
-	var callCount int64
-	wp := NewWorkerPool(
-		func(ctx context.Context, wp *WorkerPool) error {
-			atomic.AddInt64(&callCount, 1)
-			time.Sleep(10 * time.Millisecond)
-			return nil
-		},
-		WithDuration(0), // Infinite duration.
-		WithConcurrency(5),
-	)
-
-	wp.LaunchWithContext(ctx)
-	time.Sleep(50 * time.Millisecond)
-
-	// Cancel the context.
-	cancel()
-
-	// Should finish quickly.
-	done := make(chan struct{})
-	go func() {
-		wp.Wait()
-		close(done)
-	}()
-
-	select {
-	case <-done:
-		// Good, finished as expected.
-	case <-time.After(2 * time.Second):
-		t.Error("worker pool did not stop after context cancellation")
-	}
-
-	// Should have done some work.
-	if atomic.LoadInt64(&callCount) == 0 {
-		t.Error("expected some work to be done before cancellation")
-	}
-}
-
 func TestStopWithoutLaunch(t *testing.T) {
 	wp := NewWorkerPool(
 		func(ctx context.Context, wp *WorkerPool) error { return nil },
@@ -589,263 +214,5 @@ func TestWorkerStopConcurrent(t *testing.T) {
 
 	if count > 5 {
 		t.Errorf("Workers did not stop reliably, executed %d times", count)
-	}
-}
-
-func TestPauseStep(t *testing.T) {
-	callCount := atomic.Int64{}
-	workerFunc := func(ctx context.Context, wp *WorkerPool) error {
-		callCount.Add(1)
-		return nil
-	}
-
-	profile := []*Step{
-		NewStep(
-			WithConcurrency(5),
-			WithRepetitions(10),
-		),
-		NewStep(
-			WithPause(200 * time.Millisecond),
-		),
-		NewStep(
-			WithConcurrency(5),
-			WithRepetitions(10),
-		),
-	}
-
-	wp := NewMultiStepWorkerPool(workerFunc, profile)
-
-	start := time.Now()
-	wp.Launch()
-	wp.Wait()
-	duration := time.Since(start)
-
-	// Should have completed 20 calls total (10 + 10).
-	if got := callCount.Load(); got != 20 {
-		t.Errorf("expected 20 calls, got %d", got)
-	}
-
-	// Should have taken at least 200ms for the pause.
-	if duration < 200*time.Millisecond {
-		t.Errorf("expected at least 200ms duration, got %s", duration)
-	}
-}
-
-func TestLongRunningWorker(t *testing.T) {
-	workerFunc := func(ctx context.Context, wp *WorkerPool) error {
-		time.Sleep(500 * time.Millisecond) // Takes 500ms to complete.
-		return nil
-	}
-
-	wp := NewWorkerPool(workerFunc, WithRepetitions(1), WithDuration(0))
-
-	start := time.Now()
-	wp.Launch()
-	wp.Wait()
-	duration := time.Since(start)
-	// Should have taken at least 500ms.
-	if duration < 500*time.Millisecond {
-		t.Errorf("worker pool did not run for at least 500ms, got %s", duration)
-	}
-}
-
-func TestIdealTimestampsPresent(t *testing.T) {
-	// Verify that with rate-limited mode, the ideal timestamps in context are present
-	// and evenly spaced (computed as stepStart + seq/rps, sorted by value).
-	var timestamps []time.Time
-	var zeroTimeCount atomic.Int64
-	var mu sync.Mutex
-
-	workerFunc := func(ctx context.Context, wp *WorkerPool) error {
-		st := ScheduledTimeFromContext(ctx)
-		if st.IsZero() {
-			zeroTimeCount.Add(1)
-		}
-		mu.Lock()
-		timestamps = append(timestamps, st)
-		mu.Unlock()
-		return nil
-	}
-
-	wp := NewWorkerPool(
-		workerFunc,
-		WithDuration(200*time.Millisecond),
-		WithRateLimit(100, 100),
-		WithConcurrency(2),
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	if zeroTimeCount.Load() > 0 {
-		t.Errorf("expected no zero scheduled times in context, got %d", zeroTimeCount.Load())
-	}
-
-	mu.Lock()
-	defer mu.Unlock()
-
-	if len(timestamps) < 5 {
-		t.Fatalf("expected at least 5 timestamps, got %d", len(timestamps))
-	}
-
-	// Sort by timestamp value: ideal times are computed as stepStart + seq*interval,
-	// so sorted order reflects sequence assignment order.
-	sorted := make([]time.Time, len(timestamps))
-	copy(sorted, timestamps)
-	for i := 1; i < len(sorted); i++ {
-		for j := i; j > 0 && sorted[j].Before(sorted[j-1]); j-- {
-			sorted[j], sorted[j-1] = sorted[j-1], sorted[j]
-		}
-	}
-
-	// Verify timestamps are approximately 10ms apart (100 RPS = 10ms interval).
-	expectedInterval := 10 * time.Millisecond
-	tolerance := 2 * time.Millisecond
-	for i := 1; i < len(sorted); i++ {
-		gap := sorted[i].Sub(sorted[i-1])
-		if gap < expectedInterval-tolerance || gap > expectedInterval+tolerance {
-			t.Errorf("timestamp gap [%d] = %v, expected ~%v", i, gap, expectedInterval)
-		}
-	}
-}
-
-func TestUnlimitedRateNoIdealTimestamp(t *testing.T) {
-	// Verify that with RPS=0, no scheduled time appears in context.
-	var sawScheduledTime atomic.Bool
-
-	workerFunc := func(ctx context.Context, wp *WorkerPool) error {
-		st := ScheduledTimeFromContext(ctx)
-		if !st.IsZero() {
-			sawScheduledTime.Store(true)
-		}
-		return nil
-	}
-
-	wp := NewWorkerPool(
-		workerFunc,
-		WithDuration(100*time.Millisecond),
-		WithConcurrency(2),
-		// No WithRateLimit → unlimited rate.
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	if sawScheduledTime.Load() {
-		t.Error("expected no scheduled time in context for unlimited rate")
-	}
-}
-
-func TestStepTransitionsIdealTimestamps(t *testing.T) {
-	// Multi-step profile with different RPS per step. Verify both steps execute and complete.
-	var step1Calls, step2Calls atomic.Int64
-
-	step1Worker := func(ctx context.Context, wp *WorkerPool) error {
-		step1Calls.Add(1)
-		return nil
-	}
-
-	step2Worker := func(ctx context.Context, wp *WorkerPool) error {
-		step2Calls.Add(1)
-		return nil
-	}
-
-	steps := []*Step{
-		NewStep(
-			WithDuration(100*time.Millisecond),
-			WithConcurrency(2),
-			WithRateLimit(50, 50),
-			WithWorkerFunc(step1Worker),
-		),
-		NewStep(
-			WithDuration(100*time.Millisecond),
-			WithConcurrency(2),
-			WithRateLimit(100, 100),
-			WithWorkerFunc(step2Worker),
-		),
-	}
-
-	wp := NewMultiStepWorkerPool(step1Worker, steps)
-	wp.Launch()
-	wp.Wait()
-
-	if step1Calls.Load() == 0 {
-		t.Error("expected step 1 worker to be called")
-	}
-	if step2Calls.Load() == 0 {
-		t.Error("expected step 2 worker to be called")
-	}
-}
-
-func TestSlowWorkerAccumulatesLatency(t *testing.T) {
-	// Worker that sleeps 50ms with 100 RPS → later requests should have timestamps far in the past.
-	var scheduledTimes []time.Time
-	var receiptTimes []time.Time
-	var mu sync.Mutex
-
-	workerFunc := func(ctx context.Context, wp *WorkerPool) error {
-		now := time.Now()
-		st := ScheduledTimeFromContext(ctx)
-		mu.Lock()
-		scheduledTimes = append(scheduledTimes, st)
-		receiptTimes = append(receiptTimes, now)
-		mu.Unlock()
-		time.Sleep(50 * time.Millisecond) // Slow worker: 50ms per request.
-		return nil
-	}
-
-	wp := NewWorkerPool(
-		workerFunc,
-		WithDuration(300*time.Millisecond),
-		WithRateLimit(100, 100), // 100 RPS = 10ms intervals.
-		WithConcurrency(1),      // Single worker → will fall behind.
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	mu.Lock()
-	defer mu.Unlock()
-
-	if len(scheduledTimes) < 3 {
-		t.Fatalf("expected at least 3 data points, got %d", len(scheduledTimes))
-	}
-
-	// Later requests should have increasing queue time (receipt - scheduled).
-	lastQueueTime := receiptTimes[len(receiptTimes)-1].Sub(scheduledTimes[len(scheduledTimes)-1])
-	firstQueueTime := receiptTimes[0].Sub(scheduledTimes[0])
-	if lastQueueTime <= firstQueueTime {
-		t.Errorf("expected queue time to grow over time: first=%v, last=%v", firstQueueTime, lastQueueTime)
-	}
-}
-
-func TestDynamicRPSEasingIdealTimestamps(t *testing.T) {
-	// Step with RPS easing. Verify ideal timestamps are present and requests complete.
-	var callCount atomic.Int64
-	var idealTimePresent atomic.Bool
-
-	workerFunc := func(ctx context.Context, wp *WorkerPool) error {
-		if !ScheduledTimeFromContext(ctx).IsZero() {
-			idealTimePresent.Store(true)
-		}
-		callCount.Add(1)
-		return nil
-	}
-
-	wp := NewWorkerPool(
-		workerFunc,
-		WithDuration(200*time.Millisecond),
-		WithRateLimit(100, 100, EasingLinear),
-		WithConcurrency(4),
-	)
-
-	wp.Launch()
-	wp.Wait()
-
-	if callCount.Load() == 0 {
-		t.Errorf("expected at least one call with dynamic RPS easing, got %d", callCount.Load())
-	}
-	if !idealTimePresent.Load() {
-		t.Error("expected at least some calls to have a scheduled time in context")
 	}
 }
